클래스, 실행파일 이름 수정 오픈-휴무-마감 조건 제거
